V0.1.2
So I fixed this function. Now this application works on Windows. There is something else that can be done, but I will give too much unreadable code. Because if I try to change anything, it just breaks down. Especially in terms of electrons. This type of everything needs to be rewritten.