v0.1.3
OK. I again looked into this code and thought why this application did not work for me. It turns out that the author forgot or does not know that the user can only have one disk. and that's why this function was throwing an error. I just decided to run it all on another computer with three disks. and yes it works. but this is still a stupid function that works with two or more disks